Fix: MainPage의 CategorySlider 컴포넌트를 분리한다.
organism/ 하위 폴더에 컴포넌트를 별도 분리합니다.
