+++ src/components/organisms/category-slider/CategorySlider.constants.ts
