+++ src/pages/main/MainPage.tsx
@@ -1,42 +1,16 @@
-import { useEffect, useRef, useState } from "react";
-import { CgOptions as FilterIcon } from "react-icons/cg";
-import { FiArrowRight as ArrowRightIcon, FiArrowLeft as ArrowLeftIcon } from "react-icons/fi";
 import { useSelector } from "react-redux";
-import Slider from "react-slick";
 
-import { categories, initialProps, sliderSettings } from "./MainPage.constants";
-import { CategoryBox, CategoryRow, CategorySection, Divider, GridBox, GridItem, MainContainer, MainLayout, NextArrow, PortfolioSection, PrevArrow, Summary, Title, TitleSection } from "./MainPage.styled";
+import { GridBox, GridItem, MainContainer, MainLayout, PortfolioSection, Summary, Title, TitleSection } from "./MainPage.styled";
 
 import { mainPageSectionSummary } from '@/assets/data/phrase';
-import { RoundButton as FilterButton, RoundButton } from "@/components/atoms/button/Button.styled";
 import PortfolioProfile from "@/components/molecules/profile/portfolio-profile/PortfolioProfile";
+import CategorySlider from "@/components/organisms/category-slider/CategorySlider";
 import Header from "@/components/organisms/header/Header";
-import SearchModal from "@/components/organisms/modal/search-modal/SearchModal";
 import PortfolioItem from "@/components/organisms/portfolio-item/PortfolioItem";
-import useHandleSlider from "@/hooks/useHandleSlider";
 import { section } from "@/redux/sectionSlice";
 
-function Main(){
-	const [filterOpen, setFilterOpen] = useState(false);
-	const [currentCategory, setCurrentCategory] = useState('전체');
-
-	const sliderRef = useRef(null);
-	const { handlePrev, handleNext, setSlick: setSlider, currentSlideIndex } = useHandleSlider(initialProps);
-
+function MainPage(){
 	const currentSection = useSelector(section);
-
-	const openFilter = ()=> {
-		setFilterOpen((prev)=>!prev);
-	};
-
-	const handleCategory = (event: React.MouseEvent<Element, MouseEvent>)=> {
-		const eventTarget = event.target as HTMLElement;
-		setCurrentCategory(eventTarget.innerText);
-	}
-
-	useEffect(()=> {
-		setSlider(sliderRef.current!);
-	}, [])
 
 	return(
 		<MainLayout>
@@ -48,43 +22,7 @@ function Main(){
 					<Summary>{mainPageSectionSummary[currentSection]}</Summary>
 				</TitleSection>
 
-				<CategorySection>
-					<FilterButton color='Gray' onClick={openFilter}>
-						<FilterIcon size={20}/>
-						Filters
-					</FilterButton>
-
-					{ filterOpen &&
-						<SearchModal onClick={openFilter}/>
-					}
-
-					<Divider/>
-
-					<CategoryBox>
-						<PrevArrow color='White' $current={currentSlideIndex} onClick={handlePrev}>
-							<ArrowLeftIcon size={16}/>
-						</PrevArrow>
-						<NextArrow color='White' $current={currentSlideIndex} $last={11} onClick={handleNext}>
-							<ArrowRightIcon size={16}/>
-						</NextArrow>
-
-						<CategoryRow>
-							<Slider {...sliderSettings} ref={sliderRef}>
-								{categories[currentSection].map((category, index)=>{
-									let active = false;
-									if(category === currentCategory) active = true;
-									return <RoundButton
-														color='Transparency'
-														key={index}
-														onClick={handleCategory}
-														active={active}>
-														{category}
-													</RoundButton>
-								}) }
-							</Slider>
-						</CategoryRow>
-					</CategoryBox>
-				</CategorySection>
+				<CategorySlider section={currentSection}/>
 
 				<PortfolioSection>
 					<GridBox>
@@ -100,4 +38,4 @@ function Main(){
 	)
 }
 
-export default Main;
+export default MainPage;
